Applied suggested changes
Removed HowTo from naming
Updated readme files
Added verb main method
