+++ jobs/api/v3/Jobs.sln
@@ -5,14 +5,14 @@ MinimumVisualStudioVersion = 15.0.26124.0
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JobsTest", "JobsTest\JobsTest.csproj", "{FC17C0E9-2904-46E4-A197-37FB0CFD3C1D}"
 	ProjectSection(ProjectDependencies) = postProject
 		{2C5E2B0D-67F1-4DA9-98D2-F412987692C3} = {2C5E2B0D-67F1-4DA9-98D2-F412987692C3}
-		{AF57439F-4283-4813-8637-C22744FB0E52} = {AF57439F-4283-4813-8637-C22744FB0E52}
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A} = {1C28F76E-867C-4E10-8A69-9D88B091EF8A}
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuickStart", "QuickStart\QuickStart.csproj", "{129FDF6C-9E59-4D5A-B54B-A07B853E56AC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "testutil", "..\..\..\testutil\testutil.csproj", "{2C5E2B0D-67F1-4DA9-98D2-F412987692C3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HowTo", "HowTo\HowTo.csproj", "{AF57439F-4283-4813-8637-C22744FB0E52}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JobsSample", "JobsSample\JobsSample.csproj", "{1C28F76E-867C-4E10-8A69-9D88B091EF8A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -60,18 +60,18 @@ Global
 		{2C5E2B0D-67F1-4DA9-98D2-F412987692C3}.Release|x64.Build.0 = Release|Any CPU
 		{2C5E2B0D-67F1-4DA9-98D2-F412987692C3}.Release|x86.ActiveCfg = Release|Any CPU
 		{2C5E2B0D-67F1-4DA9-98D2-F412987692C3}.Release|x86.Build.0 = Release|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Debug|x64.Build.0 = Debug|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Debug|x86.Build.0 = Debug|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Release|x64.ActiveCfg = Release|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Release|x64.Build.0 = Release|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Release|x86.ActiveCfg = Release|Any CPU
-		{AF57439F-4283-4813-8637-C22744FB0E52}.Release|x86.Build.0 = Release|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Debug|x64.Build.0 = Debug|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Debug|x86.Build.0 = Debug|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Release|x64.ActiveCfg = Release|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Release|x64.Build.0 = Release|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Release|x86.ActiveCfg = Release|Any CPU
+		{1C28F76E-867C-4E10-8A69-9D88B091EF8A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
