+++ jobs/api/v3/README.md
@@ -8,21 +8,9 @@ This sample requires [.NET Core 2.0](
 
 ## Build and Run
 
-1.  **Follow the instructions in the [root README](../../README.md)**.
+* See [Quick Start](./Quickstart/README.md)
+* See [How To](./HowTo/README.md)
 
-2.  Enable APIs for your project.
-    [Click here](https://pantheon.corp.google.com/flows/enableapi?apiid=jobs.googleapis.com&showconfirmation=true)
-    to visit Cloud Platform Console and enable the Google Cloud Key Management Service API.
-
-3.  Enable data logging. 
-    [Click here](https://cloud.google.com/job-discovery/docs/before-you-begin)
-    to visit Cloud Platform Console and enable data logging..
-
-3.  From a Powershell command line:
-    ```
-    PS C:\...\dotnet-docs-samples\jobs\api\QuickStart> dotnet restore
-    PS C:\...\dotnet-docs-samples\jobs\api\QuickStart> dotnet run
-    ```
 
 ## Contributing changes
